pyproject.toml
[project]
name = "scalable-us-health-scraper"
version = "0.1.0"
description = "Scalable web scraper for US-based health leads using Playwright, Celery, and PostgreSQL/MongoDB."
authors = [
    {name = "Auto-generated",email = "auto@gen.com"}
]
requires-python = "^3.11"
dependencies = [
    "playwright (>=1.42.0,<2.0.0)",
    "celery[redis] (>=5.3.6,<6.0.0)",
    "pydantic (>=2.6.4,<3.0.0)",
    "pydantic-settings (>=2.2.1,<3.0.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "sqlalchemy (>=2.0.28,<3.0.0)",
    "motor (>=3.3.2,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "pytest (>=8.1.1,<9.0.0)",
    "pytest-asyncio (>=0.23.5,<0.24.0)",
    "pytest-playwright (>=0.4.4,<0.5.0)",
    "ollama (>=0.1.7)",
    "tweepy (>=4.14.0)",
    "pandas (>=2.0.0)",
    "openpyxl (>=3.1.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
